This compiles, not sure that it works.
StarterService trait provides the basic structure for our service.

StarterServiceServer mixes in the StarterService and extends OstrichService,
as well as implementing the neccesary start/shutdown/etc methods. Note that
some of these methods have an 'override' modifier although not needed, this is
just intended to show which methods we are implementing for the
OstrichService.

StringCodec stolen from finagle-examples, just a demo codec, not totally
interesting.

The StarterServiceConfig provides an Ostrich ServerConfig that will be used
as the basis of the config files.