Updated Client implementation to use exception handler and avoid weird sleep.
The previous implementation used an atomic counter to wait on all responses to
come back prior to calling release on the client. This was initially done as a
hack but also because the previous isReturn implementation immediately
returned false when the first error was received. Since the initial
implementation was intended to illustrate the request timeout handling, we
didn't want the client to be released until after all requests had been
submitted.

The new implementation uses the handle method to appropriately deal with a
request timeout, which eliminates the need for some kind of a counter. We also
implement a 'totalTimeout' which is the total amount of time that all requests
can take to be handled. In this case, we don't want to rescue the failure
since there is no recourse.